+++ src/CSS/Components.css
@@ -32,6 +32,7 @@ nav{
 }
 
 .page-nav {
+  animation: 0.2s ease-in 0s fadein;
   gap: 20px;
   margin-bottom: 20px;
 }
