+++ src/utils/utils.js
@@ -1,7 +1,5 @@
-import football from '../Images/football.jpg';
-import cooking from '../Images/cooking.jpg';
-import coding from '../Images/coding.jpg';
-import defaultImg from '../Images/default.jpg';
+
+
 
 export function formatDate(dateStr) {
     const created = new Date(dateStr.replace(" ", "T"))
@@ -17,15 +15,3 @@ export function buildSearchString(searchObj) {
     return string
 }
 
-export function imageSelctor(topic) {
-    switch (topic) {
-        case 'football':
-            return football;
-        case 'coding':
-            return coding;
-        case 'cooking':
-            return cooking;
-        default:
-            return defaultImg;
-    }
-}
